+++ docs/changes/2162.maintenance.md
@@ -1,1 +1,1 @@
-Move code from application scripts into modules.
+Move code from application scripts into modules to improve unit test coverage and maintainability.
